blog/requirements.txt
Django==3.2.13
Pillow==10.2.0
djangorestframework==3.15.2

flake8==6.1.0
isort==5.12.0
